requirements.txt
# ── Core LLM & Agent Framework ───────────────────────────────────────────────
langchain==0.3.25
langchain-core==0.3.60
langchain-anthropic==0.3.15
langchain-community==0.3.24
langgraph==0.4.1

# ── Vector Store & Embeddings ─────────────────────────────────────────────────
chromadb==0.6.3
sentence-transformers==3.4.1
langchain-huggingface==0.1.3

# ── Document Loading & Processing ─────────────────────────────────────────────
unstructured==0.17.2
markdown==3.7

# ── Environment & Utilities ───────────────────────────────────────────────────
python-dotenv==1.1.0
rich==13.9.4
pydantic==2.11.3
